refactor(frontend): extract PhaseCard from analysing view
Move the inline phase-row JSX from src/views/analysing.tsx
(ANALYSIS_PHASES.map at lines 1564–1660 pre-refactor) into a dedicated
PhaseCard component under src/components/analysing/phase-card.tsx,
along with the phase-card-only helpers LiveChapterRow, LiveChapterTicker,
ActivePhaseLog, LiveCastPreview, SeriesPriorPill, DroppedQuotesPanel,
HeartbeatRow, ThrottleRow, and the constants COMPLETED_PHASE_TAIL /
ACTIVE_PHASE_LOG_MAX_H / STALL_THRESHOLD_*.

Behaviour-neutral. AnalysingView shrinks from 1,769 to ~1,400 lines
and now renders `<PhaseCard {...props}/>` per phase. The 42 existing
tests in src/views/analysing.test.tsx pass unchanged — coverage of
the rendered output is preserved because PhaseCard mounts under
AnalysingView.

Opens the seam that commits 3 + 4 use for the per-phase model chip,
swap dropdown, and sticky status bar. Mirrors the listen-view
orchestrator split shipped in plan 60.
